tests: make upgrade tests run, add one for postgres
